requirements.txt
Flask==3.0.3
openai==1.65.1
python-dotenv==1.0.1
requests==2.32.3
